Allow cobrands to control which states affect reputation
Rather than assigning +1 to a user's reputation every time a report is
inspected, this allows cobrands to specify which states should increment or
decrement the reputation value when a report is inspected.

The default behaviour is for reputation to never be changed, but the Oxfordshire
cobrand will increment the reputation when a report is marked as 'action
scheduled'.

For mysociety/fixmystreetforcouncils#119